requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.10.0
Werkzeug==3.0.3
opencv-python==4.10.0.84
numpy==1.26.4
pandas>=2.2.3
openpyxl==3.1.5
reportlab==4.2.2
APScheduler==3.10.4
Pillow==10.4.0
SQLAlchemy==2.0.31
face-recognition==1.3.0
dlib==19.24.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==22.0.0
